requirements.txt
fastapi
pydantic
websocket-client